Refactored extension methods for models
This mimics what was done for entities
